GsmSST: Disregard any invalid verbose values for timezones in NITZ requests
The NITZ parser includes a special case for the emulator in which it
assumes that if a 9th field exists, it's a verbose and accurate
representation of the timezone. That field is passed along to
libcore for conversion into a TZ object, and that object is then
used, taking precedence over every other TZ representations in
the request.
libcore's getTimeZone(), however, will always return a valid timezone,
defaulting to GMT when the input is junk. If a request returns something
like "13/09/24,00:21:10+40,00,505,41132" (real example), the "+40"
will be ignored in favor of the junk "505" value, and the device ends
up in GMT instead of Sydney.
So... assume anything resulting in "GMT" is invalid if a non-zero offset
exists, so that it is used to determine the zone.

Change-Id: I361650400e71976d8327ee3929947ca26a9eff57