Remove AppOps permission check for Mms uid to bypass ICC SMS read/write issue.
With the new AppOps check added in 4.3, Icc SMS read/write is
failing when Mms app uses SMS Provider.Here the issue is that
SmsProvider runs with Phone uid and hence the package name passed
to IccSmsInterfaceManager is com.android.phone.However the calling
uid is that of Mms app since the IPC call was initiated by Mms app.
This mismatch in uid and package name causes permission denial to
ICC SMS read/write.Since the calling package name is not exposed to
SmsProvider methods, there seems no way to pass the correct package
name down to IccSmsInterfaceManager. This seems to be a fundamental
issue with the new AppOps permission check introduced in 4.3.
We are avoiding this check for Mms uid for ICC SMS read/delete
operations

Change-Id: Id165f231e35b5d64729432f7db09ba0598991b4a