Enable OMH support in Android Telephony
Support the Open Mobile Handset initative in android telephony. As per
the OMH specification in 3GPP2, Data profiles defined for specific
services like LBS, TETHERED and UNSPECIFIED are defined in the
RUIM/CSIM cards. These profiles are read from android telephony
and are used for creating data calls for the right radio technology
(CDMA/DO) when this feature is enabled.
* Supports multiple profiles for a particular OMH service type
* Supports multiple OMH profiles with same priority
* In case of OMH, for all profiles supporting SUPL, arbitration
should choose the lowest priority profile when making data calls
* Use PROFILE_TYPE_CDMA if no suitable PROFILE_TYPE_OMH available

Change-Id: I36adadfa5be42328fc3cb240626589ddde5fdb26