Update patch set 1
Patch Set 1: Code-Review-1

This patch fixes the bug discovered when running inside the emulator, but as a general policy going forward, I would much rather remove these sorts of special hacks for the Android emulator and extend or modify the Android emulator to return values that a real device would return.

In this case, the easy fix for the emulator would be to convert the user's local TZ into a quarter-hour offset, and remove the code to handle the 9th field altogether. Even better would be adding a menu or dialog to the emulator giving the developer the ability to send arbitrary NITZ time unsolicited response strings, which could be as simple as a menu item to open a single line text entry box with a generated NITZ time for the current time and GMT offset, which the user could modify or paste in a different value before sending.

Patch-set: 1
Reviewer: Jake Hamby <1003544@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1