Fix issues with SIM sms operations
Use case:
1) Boot the device with a SIM card which has some SMSs in it
2) Open the SMS app-> Settings-> Mnage SIM Card Messages
3) Options->Delete Messages

At step 3, the SMS application is crashed with reason
that the app doesn't have permissions to do a WRITE_ICC_SMS.

In SmsProvider, the calling identity is cleared with
Binder.clearCallingIdentity() for ICC operation of listing
the SIM sms and the check for permissions is done using
enforceCallingPermission(). Where as for delete/update it is
not the case and hence the issue.

Change-Id: I59bbfed2f84738d32486c85ea54bc9dafab82e7c
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ: 158326