Add empty string check for ussdMessage.
issue: ADR:ATT TS51.010 31.9.1.2
no message is displayed after USSD code *70*635*562#
Sometimes the USSD response from the network does not have any text.
That is correct and according to the spec. When the Phone App receives
this kind of response from the RIL it should replace the empty text
string with the "MMI Complete" string and display this string in the
MMIComplete window.

Change-Id: I161299843b8c19f99886868a65771d5dc1dc4846