Set language based on EFli and Efpl
Preferred language of the ME is not set according
to the preferred language stored in EFli and EFpl.

As per 3GPP TS 31.102 section 5.1.1.2, ME's preferred
language selection shall always use the EFli in preference
to the EFPL at the MF unless any of the following conditions
applies:
- if the EFLI has the value 'FFFF' in its highest priority
position, then the preferred language selection shall be the
language preference in the EFPL at the MF level according to
the procedure defined in TS 31.101
- if the ME does not support any of the language codes
indicated in EFLI, or if EFLI is not present, then the language
selection shall be as defined in EFPL at the MF level according
to the procedure defined in TS 31.101
- if neither the languages of EFLI nor EFPL are supported by the
terminal, then the terminal shall use its own internal default
selection.

This patch reads the EFli, EFpl sim files and sets the language
according to the above stated preference only if the language
is not set already.

Change-Id: Iff0bc29fad54f2950f6ad3062e9cb9fbb99ded00
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ: 122010