Do not accept incoming SMS during emergency call
When an SMS message is received while the user is making an
emergency call from the Cryptkeeper screen, all incoming SMS
messages will be lost.  This fix results in the device responding
to the network that the SMS message could not be delivered.

Explanation:
When a user makes a call from the Cryptkeeper screen there is no
broadcast receiver that can handle the incoming message. In this case
telephony will try to persist the message in the database, but as the
phone is encrypted it silently fails, while responding to the network
that the message has been received and processed.

Change-Id: Ib3f7e749b7734a5957383b03049a7a5b463915f4