WIP: telephony device controller.
component allowing the telephony framework to gain
knowledge of a representation of the hardware in order
to make educated decision on resources usage.

the telephony device controller manages a set of
hardware configuration objects.  the management sources
are: ril (real time) and overlay (initial config).

hardware base building blocks are 'modem' (which represents
a radio access technology) and 'sim' (which represents a
slot).

Change-Id: I1d486ab2fef022d41bcfcfde6895342de95e0e96
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>