SMSManager.sendMultipartTextMessage creates multiple log entries in KITKAT
This is a bug fix for a new feature of KITKAT. Android 4.4 introduces the
concept of a default SMS app. Only the default SMS app is allowed to write to
the SMS database. So code was added to the SMS dispatcher for writing the sent
SMS messages to the SMS database if the caller is not the SMS default app.
But this new code handles multipart text messages not correctly. It creates a
SMS log entry for each part instead of a single entry.

This fix adds context data to the SMSTracker so that the tracker can write a
single entry with the full text message for the last sent message part only.

Change-Id: I5b808ea098d75a621517eb714ba7d2e921eb0f85