UICC: preventing unexpected SIM state updating
UiccController requests getIccCardStatus when the radio state is changed
as "RADIO AVAILABLE" because the SIM state is not updated when the
device powered up without the card. But this enforced updating is not
needed after the radio state set as "available". In some case, this
updating makes unexpected "SIM card added" notification. This case is
occurred when the modem is restarted.

SIM card existance case:
It is not needed to monitor the radio state because the modem will send
the unsolicited events for changing the SIM state. If the modem is
restarted, the SIM card will be also restarted. Even if the radio is
restarted, it is not needed if the modem state is not changed because of
the rild restarting.

No SIM card case:
The modem can not report the card state. So, the original code is asking
the card state when the radio turns on. Even if the modem is restarted,
the card state will be "No SIM". If the card is inserted after the
radio turned on, the modem will send the changed state. If the rild
is restarted before the card insertion, the flag (mFirstRadioAvailableReceived)
is reset and the card state will be updated.

Bug: 13881264

Change-Id: Ia8054e2e5bcaea8521e8d51b62c1641509cf9f34