Introduce support for SATK Rel.9 Network search mode command
Current Network Search Mode command is part of PROVIDE LOCAL
INFORMATION SATK Rel.9 proactive command where the current
network search mode (as MANUAL or AUTOMATIC) is sent to UICC.

With is this patch, support for this command is implemented.
Note that support for this command is required for SATK Rel.9.

Change-Id: I2cecda6ea633810fd21748b7d92450a3c0ab769a
Signed-off-by: Andre Amegah <andrex.amegah@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ: 61235