Clear CallForwarding and MessageWaiting Indications during SIM Hot Swap
When SIM is removed the SIMRecords object is disposed, but
the relevant events to clear the Callforwarding and message
waiting icons are not sent.

This patch clears the call forwarding and message waiting
indicators by clearing the internal member data and sending
out notifications for the change in data.

Change-Id: Iddb5a5f28300f7f4ff519dca2f4f93ed3b201fad
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ: 64610