Introducing Emergency Call Status Intent
There could be system services which needs to know the
emergency call status before taking any drastic actions
such as shutdown of device due to excess heating.

With this patch, android.intent.action.EMERGENCY_CALL_STATUS
will be broadcasted when an emergency call is initiated and
terminated, so that system services are aware of the emergency.

Change-Id: Ie8c2c7bd388d32d48eaa15e23d85b9208a9f0a19
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ: 55480