Correct the issue while extracting mcc code
While converting the operator numeric to text string
the possibility of the operator numeric being empty thus
resulting in an exception is taken care of, but the member
variable mGotCountryCode is not correctly set. This could
result in wrong information being passed around in telephony
framework.

This patch fixes this issue by ensuring that mGotCountryCode
is set to true only after successfull conversion of the operator
numeric to text string.

Change-Id: Id08f6a1d34532da08c16c26a9238c2515ba2a2f3
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ: 198547