Fix Issue in showing network information  and signal strength
When the device doesn't support voice but supports
data & sms, and if registered to only GSM, signal
strength is not shown and also service state is shown
as No service.

Copy the Data registration state to voice registration
state if the device doesn't support voice or
voice registration state is not in service and data
registration state is in service.

Change-Id: I0ba82c4ccede22968655c1c82ef69d5065074ec5
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ: 163050