Update patch set 2
Patch Set 2: Code-Review-1

When does this happen, how can I reproduce the problem?

I'm guessing as I write this that this happens when the user changes the mode in the Settings app. If so would it be better to change the Settings app?

What are the consequences from the user perspective of not having this fixed?

Patch-set: 2
Reviewer: Wink Saville <1001401@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1