Support for OEM unsolicited communication
Android has support for unsolicited communication but doesn't
have interfaces open to java applications. There are operator
requirements for communicating with the vendor RIL which
is what this API could be used for,  e.g. most common use is for quick
test of antennas in "Tech mode" and similar apps used in production
or customer service to test devices.

Bug: 28715360
Change-Id: Ia46d723c259f3661586b8d944e625177d769f7db