Update patch set 4
Patch Set 4: Code-Review-1

Sorry, guys, but had to back away one step. getServiceProvider is actually not a simple getter - it double-checks the card and overrides spn for brand name, so, yes, the code will look a bit silly but I added a comment to make it more palatable.

I set -1 now just to get some additional verification and sign-off, but it is an important fix required from partners.

Patch-set: 4
Label: Code-Review=-1