Update patch set 1
Patch Set 1: Code-Review-1

We removed the change to hide preferred network type so this should not be needed.
It was done for bug 25012603, CL https://partner-android-review.googlesource.com/#/c/467503/

Besides, even if add this change it should be done on roaming state change and not on carrier config loaded.

Patch-set: 1
Label: Code-Review=-1