Update patch set 1
Patch Set 1:

Hi Zoran, we used Gemalto SIM reader to put type 1 email address on T-Mobile SIM card and we found X+1, X+2 bytes were written. My interpretation of the 3GPP spec is, if it is not type 1, X+1 and X+2 are mandatory. It doesn't say X+1, X+2 should not be there if it is type 1. 

If we take this change, then we might see two extra garbage characters appended at the end of email address on T-Mobile SIM.

What kind of SIM card are you using? Do you see incorrect email address imported?

One thing we might be able to do is to verify if X+1, X+2 are visible ASCIIs.

Patch-set: 1