Update patch set 4
Patch Set 4:

> Why is vold.decrypt not set? It should be set when device is
 > decrypted, and until its decrypted we anyway dont want to read the
 > SIM.

Hi Amit,

Thanks for your review.
I have discussed again with my colleague and corrected the description. The issue is that sometimes the initialization of UiccController is later than RIL reports SIM_STATUS_CHANGED. So this would cause wrong SIM status in framework. A possible solution is do a query to check SIM STATUS when initialize UiccController.

Patch-set: 4