Update patch set 4
Patch Set 4: Code-Review-1

What triggers UNSOL_RESPONSE_SIM_STATUS_CHANGED? It will lead us to query RIL_REQUEST_GET_SIM_STATUS which we do not want to do if the device is encrypted, otherwise it leads to overlapping unlock UIs in case of locked SIM (one for device unlock and one for SIM unlock). We want to wait for the device to be unlocked (decrypted) before reading sim status, which is the reason we register for radio state on if the device is not decrypted.

Please test this with an encrypted device and a locked SIM.

Patch-set: 4
Reviewer: Amit Mahajan <1054648@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1