Unreliable number(255) for voice message count should not be provided
There are SIM cards in the market that write the voice message count
0xFF to EF_MWIS when the exact number is undetermined. Since only 0x00
is handled as 'unknown' in the current Android implementation. 0xFF
should also be treated as 'unknown'.

It is better to do so from a different point of view also. 3GPP TS
23.040 says that the voice message count 255 shall be taken not only
255 but also greater number. Therefore, phone should not display the
notification "New Voicemail (255)" because the exact number of the
voice messages cannot be determined.

Bug: 30544108
Change-Id: Ic024e2063f2047b46db894d226097ab27bb6d9de