Update patch set 1
Patch Set 1:

Unfortunately, we have not found any technical document which clearly mentions this modification although there are some descriptions regarding the similar EFs such as EF MSISDN and EF FDN in 3GPP TS 27.007. We think that the same logic can be applied to EF SDN. 

  | 3GPP TS 27.007
  | 
  | 7.1 Subscriber number +CNUM
  | ...
  | if a SIM card is present or if a UICC with an active GSM application is present,
  | the information is stored in the EF MSISDN under DF Telecom.
  | If a UICC with an active USIM application is present,
  | the information is stored in the EF MSISDN under ADF USIM.
  | 
  | 7.11 Select phonebook memory storage +CPBS
  | ...
  | "FD" SIM/USIM fixdialing-phonebook.
  | If a SIM card is present or if a UICC with an active GSM application is present,
  | the information in EF FDN under DF Telecom is selected.
  | If a UICC with an active USIM application is present,
  | the information in EF FDN under ADF USIM is selected.

In addition, we applied this fix when the issue was reported with a couple of operators SIMs 3 years ago.

Patch-set: 1
Reviewer: Toshiya Ikenaga <1097598@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0