Accept multiple certification hashes in ACCF
UiccPkcs15 has a list field 'mRules' to store the contents of ACCF and a
while-loop is implemented in parseAccf(). However, only one hash data is
accepted because the second parameter 'shouldConsumeAll' of
tlvCert.parse() is true in the current implementation. It must be false
if multiple certificate hashes are acceptable.

Change-Id: Ib3e04c610ac951c44f3070e0b642551c3901e429