Update patch set 1
Patch Set 1:

AFAIK the behavior is by design. With this change registrants will be notified twice for CDMA LTE phone since that has both kinds of records. We anyway don't really have full support for CDMA phones.

Even if we want this change, we should make sure that the precise phone type is CDMA (and not CDMA LTE). What's the reason for this change though? Is it partner requested or was some failure seen?

Patch-set: 1
Label: Verified=0