Update patch set 1
Patch Set 1: Code-Review-1

Do you have an example PDU that gets mangled treating it as a string?

I agree with Christian that this is a redefinition of the wire format.  If there's evidence this is needed we'd probably need increase the protocol number to support this.

Patch-set: 1
Reviewer: Robert Greenwalt <1002609@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1